+++ example/lib/pages/subtitles_page.dart
@@ -26,6 +26,12 @@ class _SubtitlesPageState extends State<SubtitlesPage> {
     );
 
     _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
+    _betterPlayerController.addEventsListener((event) {
+      if (event.betterPlayerEventType == BetterPlayerEventType.progress) {
+        print("Current subtitle line: " +
+            _betterPlayerController.renderedSubtitle.toString());
+      }
+    });
     _setupDataSource();
     super.initState();
   }