+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.0.73
-* Added licenseUrl support for iOS DRM.
+* Added `licenseUrl` support for iOS DRM.
 * Fixed RTL text direction issue in player controls.
+* Added `renderedSubtitle` in `BetterPlayerController`.
 
 ## 0.0.72
 * Updated ExoPlayer version